+++ tests/basilisp/var_test.py
@@ -7,6 +7,7 @@ from basilisp.lang import keyword as kw
 from basilisp.lang import map as lmap
 from basilisp.lang import symbol as sym
 from basilisp.lang import vector as vec
+from basilisp.lang.exception import ExceptionInfo
 from basilisp.lang.runtime import (
     Namespace,
     NamespaceMap,
@@ -33,7 +34,7 @@ def intern_val():
     return vec.v(kw.keyword("value"))
 
 
-@pytest.fixture
+@pytest.fixture(autouse=True)
 def ns_cache(ns_sym: sym.Symbol) -> atom.Atom[NamespaceMap]:
     with patch(
         "basilisp.lang.runtime.Namespace._NAMESPACES",
@@ -43,7 +44,6 @@ def ns_cache(ns_sym: sym.Symbol) -> atom.Atom[NamespaceMap]:
 
 
 def test_public_var(
-    ns_cache: atom.Atom[NamespaceMap],
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
@@ -53,7 +53,6 @@ def test_public_var(
 
 
 def test_private_var(
-    ns_cache: atom.Atom[NamespaceMap],
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
@@ -65,7 +64,6 @@ def test_private_var(
 
 
 def test_alter_var_meta(
-    ns_cache: atom.Atom[NamespaceMap],
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
@@ -81,7 +79,6 @@ def test_alter_var_meta(
 
 
 def test_reset_var_meta(
-    ns_cache: atom.Atom[NamespaceMap],
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
@@ -96,11 +93,115 @@ def test_reset_var_meta(
     assert v.meta == lmap.m(tag=kw.keyword("async"))
 
 
+def test_var_validators(ns_sym: sym.Symbol, var_name: sym.Symbol):
+    v = Var.intern(ns_sym, var_name, 0)
+
+    even_validator = lambda i: isinstance(i, int) and i % 2 == 0
+    v.set_validator(even_validator)
+    assert even_validator == v.get_validator()
+
+    with pytest.raises(ExceptionInfo):
+        v.root = 1
+
+    with pytest.raises(ExceptionInfo):
+        v.alter_root(lambda i: i + 1)
+
+    assert 0 == v.root
+    v.root = 2
+
+    v.set_validator()
+    assert None is v.get_validator()
+
+    v.alter_root(lambda i: i + 1)
+
+    assert 3 == v.root
+
+    with pytest.raises(ExceptionInfo):
+        v.set_validator(even_validator)
+
+    odd_validator = lambda i: isinstance(i, int) and i % 2 == 1
+    v.set_validator(odd_validator)
+
+    with pytest.raises(ExceptionInfo):
+        v.root = 2
+
+
+def test_var_validators_do_fire_for_thread_local(
+    ns_sym: sym.Symbol, var_name: sym.Symbol
+):
+    v = Var.intern(ns_sym, var_name, 0, dynamic=True)
+    even_validator = lambda i: isinstance(i, int) and i % 2 == 0
+    v.set_validator(even_validator)
+
+    with pytest.raises(ExceptionInfo):
+        v.value = 5
+
+    v.value = 4
+    assert 4 == v.value
+
+
+def test_var_watchers(ns_sym: sym.Symbol, var_name: sym.Symbol):
+    v = Var.intern(ns_sym, var_name, 0)
+    assert v is v.remove_watch("nonexistent-watch")
+
+    watcher1_key = kw.keyword("watcher-the-first")
+    watcher1_vals = []
+
+    def watcher1(k, ref, old, new):
+        assert watcher1_key is k
+        assert v is ref
+        watcher1_vals.append((old, new))
+
+    v.add_watch(watcher1_key, watcher1)
+    v.alter_root(lambda v: v * 2)  # == 0
+    v.root = 4  # == 4
+
+    watcher2_key = kw.keyword("watcher-the-second")
+    watcher2_vals = []
+
+    def watcher2(k, ref, old, new):
+        assert watcher2_key is k
+        assert v is ref
+        watcher2_vals.append((old, new))
+
+    v.add_watch(watcher2_key, watcher2)
+    v.alter_root(lambda v: v * 2)  # == 8
+
+    v.remove_watch(watcher1_key)
+    v.root = 10  # == 10
+    v.alter_root(lambda v: "a" * v)  # == "aaaaaaaaaa"
+
+    assert [(0, 0), (0, 4), (4, 8)] == watcher1_vals
+    assert [(4, 8), (8, 10), (10, "aaaaaaaaaa")] == watcher2_vals
+
+
+def test_var_watchers_do_not_fire_for_thread_local(
+    ns_sym: sym.Symbol, var_name: sym.Symbol
+):
+    v = Var.intern(ns_sym, var_name, 0, dynamic=True)
+
+    watcher_vals = []
+
+    def watcher(k, ref, old, new):
+        assert "watcher" == k
+        assert v is ref
+        watcher_vals.append((old, new))
+
+    v.add_watch("watcher", watcher)
+    v.value = 10
+
+    assert not watcher_vals
+    assert 10 == v.value
+
+    v.alter_root(lambda i: i + 5)
+    assert 5 == v.root
+    assert [(0, 5)] == watcher_vals
+
+
 def test_dynamic_var(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val, dynamic=True)
     assert v.is_bound
@@ -143,7 +244,6 @@ def test_var_bindings_are_noop_for_non_dynamic_var(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val)
     assert v.is_bound
@@ -189,7 +289,6 @@ def test_intern(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val)
     assert isinstance(v, Var)
@@ -208,7 +307,8 @@ def test_intern(
 
 
 def test_intern_unbound(
-    ns_sym: sym.Symbol, var_name: sym.Symbol, ns_cache: atom.Atom[NamespaceMap]
+    ns_sym: sym.Symbol,
+    var_name: sym.Symbol,
 ):
     v = Var.intern_unbound(ns_sym, var_name)
     assert isinstance(v, Var)
@@ -227,7 +327,8 @@ def test_intern_unbound(
 
 
 def test_dynamic_unbound(
-    ns_sym: sym.Symbol, var_name: sym.Symbol, ns_cache: atom.Atom[NamespaceMap]
+    ns_sym: sym.Symbol,
+    var_name: sym.Symbol,
 ):
     v = Var.intern_unbound(ns_sym, var_name, dynamic=True)
     assert isinstance(v, Var)
@@ -263,7 +364,6 @@ def test_alter_var_root(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val)
 
@@ -286,7 +386,6 @@ def test_alter_dynamic_var_root(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val, dynamic=True)
     assert v.is_bound
@@ -329,7 +428,6 @@ def test_find_in_ns(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val)
     v_in_ns = Var.find_in_ns(ns_sym, var_name)
@@ -340,7 +438,6 @@ def test_find(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val)
     ns_qualified_sym = sym.symbol(var_name.name, ns=ns_sym.name)
@@ -352,7 +449,6 @@ def test_find_safe(
     ns_sym: sym.Symbol,
     var_name: sym.Symbol,
     intern_val,
-    ns_cache: atom.Atom[NamespaceMap],
 ):
     v = Var.intern(ns_sym, var_name, intern_val)
     ns_qualified_sym = sym.symbol(var_name.name, ns=ns_sym.name)